refactor(mobile): create useConditionalScreens bb-360
