+++ apps/mobile/src/libs/types/navigation/navigation-item.type.ts
@@ -1,12 +1,7 @@
 import { type RootNavigationParameterList } from "./root-navigation-parameter-list.type";
 
-type Properties = {
-	hasAnsweredQuizQuestions: boolean;
-	hasUser: boolean;
-};
-
 type NavigationItem = {
-	checkShouldBeRendered: (properties: Properties) => boolean;
+	shouldBeRendered: boolean;
 	component: React.ComponentType;
 	name: keyof RootNavigationParameterList;
 };
